fix: no exports of simulator should depend on jest-mock-extended
