+++ yarn-project/simulator/src/public/avm/index.ts
@@ -1,2 +1,1 @@
 export * from './avm_simulator.js';
-export * from '../fixtures/simple_contract_data_source.js';
