+++ yarn-project/simulator/src/public/index.ts
@@ -3,4 +3,3 @@ export { type PublicTxResult, PublicTxSimulator, TelemetryPublicTxSimulator } fr
 export { PublicProcessor, PublicProcessorFactory } from './public_processor/public_processor.js';
 export { PublicTxSimulationTester } from './fixtures/index.js';
 export { getCallRequestsWithCalldataByPhase } from './utils.js';
-export { SimpleContractDataSource } from './fixtures/simple_contract_data_source.js';
